+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Usage:
 ### 1.10-SNAPSHOT
 * Fix #443: Add health check enricher for SmallRye Health
 * Fix #1684: Podman builds with errors are correctly processed and reported 
+* Fix #1704: `k8s:watch` in `jkube.watch.mode=copy` doesn't work
 * Fix #1888: KubernetesExtension has helper method to add image with builder
 
 ### 1.9.1 (2022-09-14)